:lipstick: Desplazamiento de los inputs de los botones a ReferenceLampRangeUI
